defaults and overwrites
This exposes additional parameters that previousely had hardcoded
values. Additionally, this adds default values for RELEASE_DIR and
TGI_DIR, as these already had recommended values.

This changes makes it possible to run multiple servers in the same
task.

Finally, this commit exposes MODEL_PATH, as an alternative to MODEL_ID.
This allows the user to use a fully custom model without uploading it to
hf.co/